// frontend/src/models/house-options.ts
export class HouseOptions {
  water: string[] = [];
  gas: string[] = [];
  electricity: string[] = [];
  sewer: string[] = [];
  hotWater: string[] = [];
  walls: string[] = [];
  condition: string[] = [];
  bathroom: string[] = [];
  heating: string[] = [];
  bath: string[] = [];
  roof: string[] = [];
  floor: string[] = [];
}